Split WysiwygComposerTests
